refactor: move auth-context to entities layer and expose via index.ts
