+++ src/entities/auth/index.ts
@@ -1,2 +1,3 @@
 export * from './api'
 export * from './types'
+export * from './model/auth-context'
